switched enums to a separate package
