+++ core/src/main/java/bg/tu/varna/events/core/processors/businessevent/GetAllBusinessEventsProcessor.java
@@ -5,8 +5,8 @@ import bg.tu.varna.events.api.model.EventModel;
 import bg.tu.varna.events.api.operations.businessevent.getall.GetAllEventsOperation;
 import bg.tu.varna.events.api.operations.businessevent.getall.GetAllEventsRequest;
 import bg.tu.varna.events.api.operations.businessevent.getall.GetAllEventsResponse;
-import bg.tu.varna.events.persistence.entities.EventStatus;
-import bg.tu.varna.events.persistence.entities.OrganizationStatus;
+import bg.tu.varna.events.persistence.enums.EventStatus;
+import bg.tu.varna.events.persistence.enums.OrganizationStatus;
 import bg.tu.varna.events.persistence.repositories.EventRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.convert.ConversionService;
