+++ core/src/main/java/bg/tu/varna/events/core/processors/user/RefreshProcessor.java
@@ -7,8 +7,8 @@ import bg.tu.varna.events.api.operations.user.refresh.RefreshRequest;
 import bg.tu.varna.events.api.operations.user.refresh.RefreshResponse;
 import bg.tu.varna.events.core.processors.external.JwtService;
 import bg.tu.varna.events.persistence.entities.Token;
-import bg.tu.varna.events.persistence.entities.TokenType;
 import bg.tu.varna.events.persistence.entities.User;
+import bg.tu.varna.events.persistence.enums.TokenType;
 import bg.tu.varna.events.persistence.repositories.TokenRepository;
 import bg.tu.varna.events.persistence.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
