+++ core/src/main/java/bg/tu/varna/events/core/utils/ValidationUtils.java
@@ -3,7 +3,11 @@ package bg.tu.varna.events.core.utils;
 import bg.tu.varna.events.api.exceptions.OrganizationSuspendedException;
 import bg.tu.varna.events.api.exceptions.UnauthorizedActionException;
 import bg.tu.varna.events.api.exceptions.UserNotFoundException;
-import bg.tu.varna.events.persistence.entities.*;
+import bg.tu.varna.events.persistence.entities.Organization;
+import bg.tu.varna.events.persistence.entities.PersonalEvent;
+import bg.tu.varna.events.persistence.entities.Role;
+import bg.tu.varna.events.persistence.entities.User;
+import bg.tu.varna.events.persistence.enums.OrganizationStatus;
 import bg.tu.varna.events.persistence.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
