--- persistence/src/main/java/bg/tu/varna/events/persistence/entities/TokenType.java
@@ -1,6 +1,0 @@
-package bg.tu.varna.events.persistence.entities;
-
-public enum TokenType {
-	BEARER,
-	REFRESH
-}
