+++ persistence/src/main/java/bg/tu/varna/events/persistence/enums/ReservationStatus.java
@@ -1,0 +1,8 @@
+package bg.tu.varna.events.persistence.enums;
+
+public enum ReservationStatus {
+	ACTIVE,
+	CANCELED,
+	ENDED
+
+}
