+++ persistence/src/main/java/bg/tu/varna/events/persistence/enums/TokenType.java
@@ -1,0 +1,6 @@
+package bg.tu.varna.events.persistence.enums;
+
+public enum TokenType {
+	BEARER,
+	REFRESH
+}
